+++ Changelog.md
@@ -5,7 +5,7 @@
 * CXV1-9007 - Hide interaction bar based on number of interactions
 * CXV1-9014 - Added side panel styling, hide side panel when no script, and added button to open side panel when collapsed in toolbarMode
 * CXV1-9013 - Toolbar view of custom fields && add custom fields to email interactions
-* CXV1-10710 - Interaction bar overflow
+* CXV1-10710 - Interaction bar overflow shadows and notifications for unresponded messages and pending interactions
 * CXV1-10156 - Trigger end interaction when agent ends chat before receiving customer reply
 * CXV1-9012 - Transfer controls UI updates for Toolbar and AgentDesktop